refactor(perf): move benchmark logic into Makefile, split bench file
Address review feedback from @fseldow:
- Makefile now owns the benchmark logic via shareable variables
  (BENCH_HEAD, BENCH_BASE, BENCH_COUNT, BENCH_THRESHOLD, BENCH_BASE_REF)
  and targets: benchmark, benchmark-base, benchmark-gate, benchmark-ci.
  This makes local reproduction and troubleshooting easy.
- benchmark.yml: each CI step now just calls a make target; dropped the
  hardcoded base.txt/head.txt and the redundant pull_request guard on
  the head run; still uploads results on failure for troubleshooting.
- benchmark-gate.sh: auto-install a pinned benchstat when missing
  instead of exiting, so the gate works out of the box.
- Rename result files to self-documenting benchmark-head.txt /
  benchmark-base.txt; .gitignore now uses benchmark*.txt.
- Split the executor benchmarks into executor_benchmark_test.go; the
  shared sync.Once mock registration stays in executor_test.go.
